# src/main/resources/openapi/marketplace-onboarding-api.yml
openapi: 3.0.3
info:
  title: API Marketplace Onboarding
  description: Accepts inbound requests for API onboarding, publishing, new API proposals, and general contact.
  version: 1.0.0
servers:
  - url: /v1
paths:
  /requests/onboarding:
    post:
      summary: Submit an API onboarding request
      description: Request access to an existing API in the marketplace.
      operationId: createOnboardingRequest
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OnboardingRequest'
            example:
              name: Colin Smith
              organisation: HMCTS
              email: colin@example.com
              jobTitle: Software Engineer
              phone: "07700900000"
              apiRequested: RAG Service API
              environment: sandbox
              callVolume: 1000/day
              useCase: Automating hearing result document distribution
      responses:
        '201':
          description: Request accepted
        '400':
          description: Invalid request body
...
components:
...
  schemas:
    OnboardingRequest:
      type: object
      properties:
        name:
          type: string
          description: Full name of the requester
        organisation:
          type: string
          description: Organisation the requester belongs to
        email:
          type: string
          format: email
        jobTitle:
          type: string
        phone:
          type: string
        apiRequested:
          type: string
          description: Name of the API being requested
        environment:
          type: string
          description: Target environment (e.g. sandbox, dev, prod)
        callVolume:
          type: string
          description: Expected call volume (e.g. 1000/day)
        useCase:
          type: string
          description: Description of the intended use case
...